Added code for modifying the data from front end

+++ src/sample/ProductTable.java
@@ -33,6 +33,7 @@ import java.util.ResourceBundle;
 public class ProductTable implements Initializable {
 
     public static Stage insertproduct;
+    public static Product p;
 
     @FXML
     TableView<Product> tableproduct;
@@ -268,56 +269,57 @@ public class ProductTable implements Initializable {
                 //TimeUnit.SECONDS.sleep(3);
                 al.close();
             }
-            catch (SQLException sq)
-            {   tableproduct.getItems().clear();
+            catch (SQLException sq) {
+                try{
+                tableproduct.getItems().clear();
                 ResultSet result = st.executeQuery("select * from product");
                 System.out.println("F");
                 while (result.next()) {   //System.out.println(rs.next());
-                    Integer a=Integer.parseInt(result.getString("pid"));
+                    Integer a = Integer.parseInt(result.getString("pid"));
                     //System.out.println("C");
-                    String b=result.getString("pname");
+                    String b = result.getString("pname");
                     //System.out.println("D");
-                    Integer c=Integer.parseInt(result.getString("pcost"));
+                    Integer c = Integer.parseInt(result.getString("pcost"));
                     //System.out.println("E");
-                    Integer d=Integer.parseInt(result.getString("profit"));
+                    Integer d = Integer.parseInt(result.getString("profit"));
                     //System.out.println("G");
-                    Integer e=Integer.parseInt(result.getString("catid"));
-                    tableproduct.getItems().add(new Product(a,b,c,d,e));
+                    Integer e = Integer.parseInt(result.getString("catid"));
+                    tableproduct.getItems().add(new Product(a, b, c, d, e));
 
                 }
-                StringBuilder str=new StringBuilder("");
+                StringBuilder str = new StringBuilder("");
                 Class.forName("com.mysql.cj.jdbc.Driver");
                 //String url = "jdbc:mysql://localhost:3306/stud?allowPublicKeyRetrieval=true&useSSL=false";
                 //Connection con = DriverManager.getConnection(url,"root","sc13111998");
                 //Statement st=con.createStatement();
-                ResultSet rs=st.executeQuery(cmd.getText());
-                ResultSetMetaData rsm=rs.getMetaData();
-                String s=DBTablePrinter.printResultSet(rs);
+                ResultSet rs = st.executeQuery(cmd.getText());
+                ResultSetMetaData rsm = rs.getMetaData();
+                String s = DBTablePrinter.printResultSet(rs);
                 System.out.println(rsm);
                 System.out.println(rsm.getColumnCount());
                 for (int i = 1; i <= rsm.getColumnCount(); i++) {
-                    str.append(rsm.getColumnName(i)+"           ");
+                    str.append(rsm.getColumnName(i) + "           ");
                 }
                 str.append("\n----------------------------------------------------------------------------------------------" +
                         "---------------------------------------------------------------------------------------------------\n");
                 while (rs.next()) {
                     for (int i = 1; i <= rsm.getColumnCount(); i++) {
-                        str.append(rs.getString(i)+"            ");
+                        str.append(rs.getString(i) + "            ");
                     }
                     str.append("\n");
                 }
-                JFXDialogLayout content=new JFXDialogLayout();
-                Label label=new Label(" OUTPUT ");
+                JFXDialogLayout content = new JFXDialogLayout();
+                Label label = new Label(" OUTPUT ");
                 label.setStyle("-fx-text-fill:#fff;-fx-font-weight:bold;-fx-font-size:30px;-fx-alignment:center;-fx-font-family:Lato;-fx-border-color:#fff;-fx-border-width:4px;-fx-border-radius:10px;");
                 label.setAlignment(Pos.CENTER);
                 content.setHeading(label);
-                TextArea textArea=new TextArea(s);
+                TextArea textArea = new TextArea(s);
                 textArea.setStyle("-fx-font-weight:bold;");
                 content.setBody(textArea);
-                JFXDialog dialog=new JFXDialog(stackpane,content,JFXDialog.DialogTransition.BOTTOM);
+                JFXDialog dialog = new JFXDialog(stackpane, content, JFXDialog.DialogTransition.BOTTOM);
                 content.setStyle("-fx-background-color:#2ECC71;-fx-pref-width:600px;-fx-pref-height:450px;-fx-text-fill:#ff0000;-fx-text-color:#ff0000;");
                 dialog.setContent(content);
-                JFXButton button=new JFXButton("Okay");
+                JFXButton button = new JFXButton("Okay");
                 button.setStyle("-fx-background-color:#303030;-fx-text-fill:#fff;-fx-font-weight:bold;-fx-pref-width:150px;-fx-pref-height:40px;-fx-background-radius:20px;-fx-border-radius:20px;");
                 button.setOnAction(new EventHandler<ActionEvent>() {
                     @Override
@@ -327,6 +329,16 @@ public class ProductTable implements Initializable {
                 });
                 content.setActions(button);
                 dialog.show();
+            }catch (Exception e)
+            {
+                Alert al = new Alert(Alert.AlertType.ERROR);
+                al.setTitle("OOPS!!!");
+                al.setHeaderText(null);
+                al.setContentText("WRONG INPUT!!!");
+                al.showAndWait();
+                //TimeUnit.SECONDS.sleep(3);
+                al.close();
+            }
             }
 
 
@@ -337,6 +349,7 @@ public class ProductTable implements Initializable {
         cmd.setText("");
     }
     public void add(ActionEvent event) throws IOException {
+        p=null;
         insertproduct = new Stage();
         Parent root = FXMLLoader.load(getClass().getResource("FXML/Insertproduct1.fxml"));
         insertproduct.setTitle("New Product");
@@ -345,36 +358,65 @@ public class ProductTable implements Initializable {
         insertproduct.setScene(new Scene(root, 860, 609));
         insertproduct.show();
     }
+    public void modify(ActionEvent event) throws IOException{
+        p=tableproduct.getSelectionModel().getSelectedItem();
+        if(p!=null) {
+            insertproduct = new Stage();
+            Parent root = FXMLLoader.load(getClass().getResource("FXML/Insertproduct1.fxml"));
+            insertproduct.setTitle("Modify Product");
+            insertproduct.initStyle(StageStyle.UNDECORATED);
+            insertproduct.setResizable(false);
+            insertproduct.setScene(new Scene(root, 860, 609));
+            insertproduct.show();
+        }
+        else{
+            Alert al = new Alert(Alert.AlertType.ERROR);
+            al.setTitle("OOPS!!!");
+            al.setHeaderText(null);
+            al.setContentText("Choose a record to modify");
+            al.showAndWait();
+            //TimeUnit.SECONDS.sleep(3);
+            al.close();
+        }
+    }
     public void deletesel(ActionEvent event) throws Exception {
         System.out.println("Shubham Chaudhary");
         Product p=tableproduct.getSelectionModel().getSelectedItem();
-        String url = "jdbc:mysql://localhost:3306/galleria?useSSL=false";
-        String uname = "root";
-        String pass = "sc13111998";
-        Class.forName("com.mysql.cj.jdbc.Driver");
-        Connection con = DriverManager.getConnection(url, uname, pass);
-        Statement st = con.createStatement();
-        Alert al = new Alert(Alert.AlertType.CONFIRMATION);
-        al.setTitle("ALERT!!!");
-        al.setHeaderText(null);
-        al.setContentText("Sure u want to delete??");
-        Optional<ButtonType> op=al.showAndWait();
-        //TimeUnit.SECONDS.sleep(3);
-        try {
-            if(op.get()==ButtonType.OK && op.isPresent()) {
-                System.out.println("1");
-                st.executeUpdate("delete from product where pid=" + p.getPid() + ";");
-                addfromdb();
-                //System.out.println(c.getCid() + " " + c.getCname() + " " + c.getEmail() + " " + c.getPhno());
-            }
-            else {
-                System.out.println("2");
-                al.close();
+        if(p!=null) {
+            String url = "jdbc:mysql://localhost:3306/galleria?useSSL=false";
+            String uname = "root";
+            String pass = "sc13111998";
+            Class.forName("com.mysql.cj.jdbc.Driver");
+            Connection con = DriverManager.getConnection(url, uname, pass);
+            Statement st = con.createStatement();
+            Alert al = new Alert(Alert.AlertType.CONFIRMATION);
+            al.setTitle("ALERT!!!");
+            al.setHeaderText(null);
+            al.setContentText("Sure u want to delete??");
+            Optional<ButtonType> op = al.showAndWait();
+            //TimeUnit.SECONDS.sleep(3);
+            try {
+                if (op.get() == ButtonType.OK && op.isPresent()) {
+                    System.out.println("1");
+                    st.executeUpdate("delete from product where pid=" + p.getPid() + ";");
+                    addfromdb();
+                    //System.out.println(c.getCid() + " " + c.getCname() + " " + c.getEmail() + " " + c.getPhno());
+                } else {
+                    System.out.println("2");
+                    al.close();
+                }
+            } catch (Exception ex) {
+                System.out.println(ex);
             }
         }
-        catch (Exception ex)
-        {
-            System.out.println(ex);
+        else {
+            Alert al = new Alert(Alert.AlertType.ERROR);
+            al.setTitle("OOPS!!!");
+            al.setHeaderText(null);
+            al.setContentText("Choose a record to delete");
+            al.showAndWait();
+            //TimeUnit.SECONDS.sleep(3);
+            al.close();
         }
